+++ vk_sdk/__init__.py
@@ -8,8 +8,8 @@
 from .lp import BotLongPoll, UserLongPoll
 from .user import User
 from .keyboard import Keyboard
-from .cmd import start_command, command, after_text_matcher, after_func
+from .cmd import menu, command, after_text_matcher, after_func
 
 __author__ = 'SPRAVEDLIVO'
-__version__ = '1.6.3'
+__version__ = '1.7'
 __email__ = 'admin@spravedlivo.dev'