+++ vk_sdk/imports.py
@@ -1,3 +1,5 @@
+"""Python importing simplified!"""
+
 import os
 from importlib import util
 import importlib